requirements.txt
# SingllLive - 程序员深夜电台 依赖包

# B站弹幕相关
# blivedm 从 GitHub 安装 (PyPI 上的 "blivedm" 是不同的包)
git+https://github.com/xfgryujk/blivedm.git@master
bilibili-api-python>=0.15.0

# HTTP 和异步
aiohttp>=3.9.0

# OBS WebSocket v5 控制 (OBS Studio 28+ 内置)
obsws-python>=1.7.0

# 图像处理
Pillow>=9.0.0
